Manager and client database changes

+++ src/main/java/com/taw/polybank/entity/BankAccountEntity.java
@@ -22,7 +22,7 @@ public class BankAccountEntity {
     private byte active;
     @Basic
     @Column(name = "balance")
-    private float balance;
+    private double balance;
     @Basic
     @Column(name = "Badge_id")
     private int badgeId;
@@ -59,7 +59,7 @@ public class BankAccountEntity {
         this.active = active;
     }
 
-    public float getBalance() {
+    public double getBalance() {
         return balance;
     }
